abstract sargable parser interface
